requirements/base.txt
django-allauth==0.33.0
django-bootstrap4==0.0.1
django-environ==0.4.4
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
git+https://github.com/Lamelos/django-allauth-office365.git
django-admin-view-permission
django-reversion==2.0.12